# build/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "@infavy/api-server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/home/similar": {
      "get": {
        "summary": "Get similar videos for a given video",
        "parameters": [
          {
            "in": "query",
            "name": "excludeVideoId",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/home/video/{id}": {
      "get": {
        "summary": "Get a video by id or query params",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/profiles/{uid}": {
      "get": {
        "summary": "Get profiles for a user",
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
...
}